+++ package.json
@@ -24,6 +24,7 @@
     "eth-json-rpc-filters": "^4.2.1",
     "ethereumjs-tx": "^2.1.1",
     "ganache-cli": "^6.10.2",
+    "mocha-parallel-tests": "^2.3.0",
     "node-fetch-polyfill": "^2.0.6",
     "truffle": "^5.1.46",
     "web3": "^1.3.0"
